--- lib/time.ts
@@ -1,7 +1,0 @@
-export const convertMinsToHrsMins = (mins) => {
-  let h = Math.floor(mins / 60);
-  let m = mins % 60;
-  h = h < 10 ? "0" + h : h;
-  m = m < 10 ? "0" + m : m;
-  return `${h}:${m}`;
-};
